GraphQL: extracting traits into new package
