+++ app/models/GQLSchema.scala
@@ -1,459 +1,13 @@
 package models
 
-import models.entities.DNA.{Gene, SimpleVariant, Variant}
+import models.entities.DNA.SimpleVariant
 import models.entities.Entities._
 import models.Functions._
-import models.GQLSchema.{gene, genesFetcher, scoredGene, studiesFetcher, study, variant, variantsFetcher}
 import models.entities.DNA
+import models.gql.{GQLArguments, GQLGene, GQLIndexVariantAssociation, GQLManhattanAssociation, GQLStudy, GQLStudyLeadVariantAssociation, GQLTagVariantAssociation, GQLVariant}
 import sangria.execution.deferred._
 import sangria.macros.derive._
-import sangria.schema
 import sangria.schema.{Field, _}
-
-trait GQLGene {
-  implicit val geneHasId = HasId[Gene, String](_.id)
-
-  val genesFetcher = Fetcher(
-    config = FetcherConfig.maxBatchSize(100),
-    fetch = (ctx: Backend, geneIds: Seq[String]) => {
-      ctx.getGenes(geneIds)
-    })
-
-  implicit val gene = deriveObjectType[Backend, Gene]()
-
-  val scoredGene = ObjectType(
-    "ScoredGene",
-    "This object link a Gene with a score",
-    fields[Backend, (String, Double)](
-      Field("gene", gene, Some("Gene Info"), resolve = rsl => genesFetcher.defer(rsl.value._1)),
-      Field(
-        "score",
-        FloatType,
-        Some("Score a Float number between [0. .. 1.]"),
-        resolve = _.value._2)))
-
-}
-
-trait GQLVariant {
-  implicit val variantHasId = HasId[Variant, String](_.id)
-
-  val variantsFetcher = Fetcher(
-    config = FetcherConfig.maxBatchSize(1000),
-    fetch = (ctx: Backend, variantIds: Seq[String]) => {
-      ctx.getVariants(variantIds)
-    })
-
-  val variant = deriveObjectType[Backend, Variant](
-    ExcludeFields("caddAnnotation", "gnomadAnnotation", "annotation"),
-    DocumentField("rsId", "Approved symbol name of a gene"),
-    DocumentField("chromosome", "Ensembl Gene ID of a gene"),
-    DocumentField("position", "Approved symbol name of a gene"),
-    DocumentField("chromosomeB37", "chrom ID GRCH37"),
-    DocumentField("positionB37", "Approved symbol name of a gene"),
-    AddFields(
-      Field(
-        "nearestGene",
-        OptionType(gene),
-        Some("Nearest gene"),
-        resolve = el => genesFetcher.deferOpt(el.value.annotation.nearestGeneId)),
-      Field(
-        "nearestGeneDistance",
-        OptionType(LongType),
-        Some("Distance to the nearest gene (any biotype)"),
-        resolve = _.value.annotation.nearestGeneDistance),
-      Field(
-        "nearestCodingGene",
-        OptionType(gene),
-        Some("Nearest protein-coding gene"),
-        resolve = el => genesFetcher.deferOpt(el.value.annotation.nearestCodingGeneId)),
-      Field(
-        "nearestCodingGeneDistance",
-        OptionType(LongType),
-        Some("Distance to the nearest gene (protein-coding biotype)"),
-        resolve = _.value.annotation.nearestCodingGeneDistance),
-      Field(
-        "mostSevereConsequence",
-        OptionType(StringType),
-        Some("Most severe consequence"),
-        resolve = _.value.annotation.mostSevereConsequence),
-      Field(
-        "caddRaw",
-        OptionType(FloatType),
-        Some("Combined Annotation Dependent Depletion - Raw score"),
-        resolve = _.value.caddAnnotation.raw),
-      Field(
-        "caddPhred",
-        OptionType(FloatType),
-        Some("Combined Annotation Dependent Depletion - Scaled score"),
-        resolve = _.value.caddAnnotation.phred),
-      Field(
-        "gnomadAFR",
-        OptionType(FloatType),
-        Some("gnomAD Allele frequency (African/African-American population)"),
-        resolve = _.value.gnomadAnnotation.afr),
-      Field(
-        "gnomadAMR",
-        OptionType(FloatType),
-        Some("gnomAD Allele frequency (Latino/Admixed American population)"),
-        resolve = _.value.gnomadAnnotation.amr),
-      Field(
-        "gnomadASJ",
-        OptionType(FloatType),
-        Some("gnomAD Allele frequency (Ashkenazi Jewish population)"),
-        resolve = _.value.gnomadAnnotation.asj),
-      Field(
-        "gnomadEAS",
-        OptionType(FloatType),
-        Some("gnomAD Allele frequency (East Asian population)"),
-        resolve = _.value.gnomadAnnotation.eas),
-      Field(
-        "gnomadFIN",
-        OptionType(FloatType),
-        Some("gnomAD Allele frequency (Finnish population)"),
-        resolve = _.value.gnomadAnnotation.fin),
-      Field(
-        "gnomadNFE",
-        OptionType(FloatType),
-        Some("gnomAD Allele frequency (Non-Finnish European population)"),
-        resolve = _.value.gnomadAnnotation.nfe),
-      Field(
-        "gnomadNFEEST",
-        OptionType(FloatType),
-        Some("gnomAD Allele frequency (Non-Finnish Eurpoean Estonian sub-population)"),
-        resolve = _.value.gnomadAnnotation.nfeEST),
-      Field(
-        "gnomadNFENWE",
-        OptionType(FloatType),
-        Some(
-          "gnomAD Allele frequency (Non-Finnish Eurpoean North-Western European sub-population)"),
-        resolve = _.value.gnomadAnnotation.nfeNWE),
-      Field(
-        "gnomadNFESEU",
-        OptionType(FloatType),
-        Some("gnomAD Allele frequency (Non-Finnish Eurpoean Southern European sub-population)"),
-        resolve = _.value.gnomadAnnotation.nfeSEU),
-      Field(
-        "gnomadNFEONF",
-        OptionType(FloatType),
-        Some(
-          "gnomAD Allele frequency (Non-Finnish Eurpoean Other non-Finnish European sub-population)"),
-        resolve = _.value.gnomadAnnotation.nfeONF),
-      Field(
-        "gnomadOTH",
-        OptionType(FloatType),
-        Some("gnomAD Allele frequency (Other (population not assigned) population)"),
-        resolve = _.value.gnomadAnnotation.oth)))
-
-}
-
-trait GQLStudy {
-  implicit val studyHasId = HasId[Study, String](_.studyId)
-
-  val studiesFetcher = Fetcher(
-    config = FetcherConfig.maxBatchSize(100),
-    fetch = (ctx: Backend, stids: Seq[String]) => {
-      ctx.getStudies(stids)
-    })
-
-  val study = deriveObjectType[Backend, Study](
-    DocumentField("traitReported", "Trait Label as reported on the publication"),
-    DocumentField("traitEfos", "A list of curated efo codes"),
-    DocumentField("pubId", "PubMed ID for the corresponding publication"),
-    DocumentField("pubDate", "Publication Date as YYYY-MM-DD"),
-    DocumentField("pubJournal", "Publication Journal name"),
-    DocumentField("hasSumstats", "Contains summary statistical information"),
-    AddFields(
-      Field(
-        "nTotal",
-        LongType,
-        Some("n total cases (n initial + n replication)"),
-        resolve = r => r.value.nInitial.getOrElse(0L) + r.value.nReplication.getOrElse(0L))))
-
-}
-
-trait GQLIndexVariantAssociation {
-
-  val indexVariantAssociation = ObjectType(
-    "IndexVariantAssociation",
-    "This object represent a link between a triple (study, trait, index_variant) and a tag variant " +
-      "via an expansion method (either ldExpansion or FineMapping)",
-    fields[Backend, V2DRow](
-      Field(
-        "tagVariant",
-        variant,
-        Some("Tag variant ID as ex. 1_12345_A_T"),
-        resolve = rsl => variantsFetcher.defer(rsl.value.tag.id)),
-      Field(
-        "study",
-        study,
-        Some("study ID"),
-        resolve = rsl => studiesFetcher.defer(rsl.value.study.studyId)),
-      Field(
-        "pval",
-        FloatType,
-        Some("p-val between a study and an the provided index variant"),
-        resolve =
-          r => toSafeDouble(r.value.association.pvalMantissa, r.value.association.pvalExponent)
-      ), // TODO TEMPORAL HACK
-      Field(
-        "pvalMantissa",
-        FloatType,
-        Some("p-val between a study and an the provided index variant"),
-        resolve = _.value.association.pvalMantissa),
-      Field(
-        "pvalExponent",
-        LongType,
-        Some("p-val between a study and an the provided index variant"),
-        resolve = _.value.association.pvalExponent),
-      Field(
-        "nTotal",
-        LongType,
-        Some("n total cases (n initial + n replication)"),
-        resolve =
-          r => r.value.study.nInitial.getOrElse(0L) + r.value.study.nReplication.getOrElse(0L)),
-      Field("nCases", LongType, Some("n cases"), resolve = _.value.study.nCases.getOrElse(0L)),
-      Field("overallR2", OptionType(FloatType), Some("study ID"), resolve = _.value.association.r2),
-      Field(
-        "afr1000GProp",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.association.afr1000GProp),
-      Field(
-        "amr1000GProp",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.association.amr1000GProp),
-      Field(
-        "eas1000GProp",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.association.eas1000GProp),
-      Field(
-        "eur1000GProp",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.association.eur1000GProp),
-      Field(
-        "sas1000GProp",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.association.sas1000GProp),
-      Field("log10Abf", OptionType(FloatType), Some(""), resolve = _.value.association.log10Abf),
-      Field(
-        "posteriorProbability",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.association.posteriorProbability),
-      Field("oddsRatio", OptionType(FloatType), Some(""), resolve = _.value.odds.oddsCI),
-      Field(
-        "oddsRatioCILower",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.odds.oddsCILower),
-      Field(
-        "oddsRatioCIUpper",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.odds.oddsCIUpper),
-      Field("beta", OptionType(FloatType), Some(""), resolve = _.value.beta.betaCI),
-      Field("betaCILower", OptionType(FloatType), Some(""), resolve = _.value.beta.betaCILower),
-      Field("betaCIUpper", OptionType(FloatType), Some(""), resolve = _.value.beta.betaCIUpper),
-      Field("direction", OptionType(StringType), Some(""), resolve = _.value.beta.direction)))
-
-}
-
-trait GQLStudyLeadVariantAssociation {
-
-  val studiesAndLeadVariantsForGene = ObjectType(
-    "StudiesAndLeadVariantsForGene",
-    "A list of Studies and Lead Variants for a Gene",
-    fields[Backend, LeadRow](
-      Field(
-        "indexVariant",
-        variant,
-        Some("Tag variant ID as ex. 1_12345_A_T"),
-        resolve = rsl => variantsFetcher.defer(rsl.value.leadVariant.id)),
-      Field(
-        "study",
-        study,
-        Some("study ID"),
-        resolve = rsl => studiesFetcher.defer(rsl.value.studyId)),
-      Field(
-        "pval",
-        FloatType,
-        Some("p-val between a study and an the provided index variant"),
-        resolve = r => toSafeDouble(r.value.pvalMantissa, r.value.pvalExponent)
-      ), // TODO TEMPORAL HACK
-      Field(
-        "pvalMantissa",
-        FloatType,
-        Some("p-val between a study and an the provided index variant"),
-        resolve = _.value.pvalMantissa),
-      Field(
-        "pvalExponent",
-        LongType,
-        Some("p-val between a study and an the provided index variant"),
-        resolve = _.value.pvalExponent),
-      Field("oddsRatio", OptionType(FloatType), Some(""), resolve = _.value.odds.oddsCI),
-      Field(
-        "oddsRatioCILower",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.odds.oddsCILower),
-      Field(
-        "oddsRatioCIUpper",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.odds.oddsCIUpper),
-      Field("beta", OptionType(FloatType), Some(""), resolve = _.value.beta.betaCI),
-      Field("betaCILower", OptionType(FloatType), Some(""), resolve = _.value.beta.betaCILower),
-      Field("betaCIUpper", OptionType(FloatType), Some(""), resolve = _.value.beta.betaCIUpper),
-      Field("direction", OptionType(StringType), Some(""), resolve = _.value.beta.direction)))
-
-}
-
-trait GQLTagVariantAssociation {
-
-  val tagVariantAssociation = ObjectType(
-    "TagVariantAssociation",
-    "This object represent a link between a triple (study, trait, index_variant) and a tag variant " +
-      "via an expansion method (either ldExpansion or FineMapping)",
-    fields[Backend, V2DRow](
-      Field(
-        "indexVariant",
-        variant,
-        Some("Tag variant ID as ex. 1_12345_A_T"),
-        resolve = rsl => variantsFetcher.defer(rsl.value.lead.id)),
-      Field(
-        "study",
-        study,
-        Some("study ID"),
-        resolve = rsl => studiesFetcher.defer(rsl.value.study.studyId)),
-      Field(
-        "pval",
-        FloatType,
-        Some("p-val between a study and an the provided index variant"),
-        resolve =
-          r => toSafeDouble(r.value.association.pvalMantissa, r.value.association.pvalExponent)
-      ), // TODO TEMPORAL HACK
-      Field(
-        "pvalMantissa",
-        FloatType,
-        Some("p-val between a study and an the provided index variant"),
-        resolve = _.value.association.pvalMantissa),
-      Field(
-        "pvalExponent",
-        LongType,
-        Some("p-val between a study and an the provided index variant"),
-        resolve = _.value.association.pvalExponent),
-      Field(
-        "nTotal",
-        LongType,
-        Some("n total cases (n initial + n replication)"),
-        resolve =
-          r => r.value.study.nInitial.getOrElse(0L) + r.value.study.nReplication.getOrElse(0L)),
-      Field("nCases", LongType, Some("n cases"), resolve = _.value.study.nCases.getOrElse(0L)),
-      Field("overallR2", OptionType(FloatType), Some("study ID"), resolve = _.value.association.r2),
-      Field(
-        "afr1000GProp",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.association.afr1000GProp),
-      Field(
-        "amr1000GProp",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.association.amr1000GProp),
-      Field(
-        "eas1000GProp",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.association.eas1000GProp),
-      Field(
-        "eur1000GProp",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.association.eur1000GProp),
-      Field(
-        "sas1000GProp",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.association.sas1000GProp),
-      Field("log10Abf", OptionType(FloatType), Some(""), resolve = _.value.association.log10Abf),
-      Field(
-        "posteriorProbability",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.association.posteriorProbability),
-      Field("oddsRatio", OptionType(FloatType), Some(""), resolve = _.value.odds.oddsCI),
-      Field(
-        "oddsRatioCILower",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.odds.oddsCILower),
-      Field(
-        "oddsRatioCIUpper",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.odds.oddsCIUpper),
-      Field("beta", OptionType(FloatType), Some(""), resolve = _.value.beta.betaCI),
-      Field("betaCILower", OptionType(FloatType), Some(""), resolve = _.value.beta.betaCILower),
-      Field("betaCIUpper", OptionType(FloatType), Some(""), resolve = _.value.beta.betaCIUpper),
-      Field("direction", OptionType(StringType), Some(""), resolve = _.value.beta.direction)))
-
-}
-
-trait GQLManhattanAssociation extends GQLGene {
-
-  val manhattanAssociation = deriveObjectType[Backend, ManhattanAssociation](
-    ExcludeFields("v2dOdds", "v2dBeta", "variantId", "pval", "bestGenes", "bestColocGenes", "bestL2Genes"),
-    DocumentField("credibleSetSize", "The cardinal of the set defined as tag variants for an index variant coming from crediblesets"),
-    DocumentField("ldSetSize", "The cardinal of the set defined as tag variants for an index variant coming from ld expansion"),
-    DocumentField("totalSetSize", "The cardinal of the set defined as tag variants for an index variant coming from any expansion"),
-    AddFields(
-      Field(
-        "variant",
-        variant,
-        Some("Index variant"),
-        resolve = r => variantsFetcher.defer(r.value.variantId)),
-      Field(
-        "pval",
-        FloatType,
-        Some("Computed p-Value"),
-        resolve = r => toSafeDouble(r.value.pvalMantissa, r.value.pvalExponent)),
-      Field("oddsRatio", OptionType(FloatType), Some(""), resolve = _.value.v2dOdds.oddsCI),
-      Field(
-        "oddsRatioCILower",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.v2dOdds.oddsCILower),
-      Field(
-        "oddsRatioCIUpper",
-        OptionType(FloatType),
-        Some(""),
-        resolve = _.value.v2dOdds.oddsCIUpper),
-      Field("beta", OptionType(FloatType), Some(""), resolve = _.value.v2dBeta.betaCI),
-      Field("betaCILower", OptionType(FloatType), Some(""), resolve = _.value.v2dBeta.betaCILower),
-      Field("betaCIUpper", OptionType(FloatType), Some(""), resolve = _.value.v2dBeta.betaCIUpper),
-      Field("direction", OptionType(StringType), Some(""), resolve = _.value.v2dBeta.direction),
-      Field(
-        "bestGenes",
-        ListType(scoredGene),
-        Some("A list of best genes associated"),
-        resolve = _.value.bestGenes),
-      Field(
-        "bestColocGenes",
-        ListType(scoredGene),
-        Some("A list of best genes associated"),
-        resolve = _.value.bestColocGenes),
-      Field(
-        "bestLocus2Genes",
-        ListType(scoredGene),
-        Some("A list of best L2G scored genes associated"),
-        resolve = _.value.bestL2Genes))
-  )
-
-}
 
 object GQLSchema
   extends GQLGene
@@ -462,50 +16,8 @@ object GQLSchema
     with GQLIndexVariantAssociation
     with GQLTagVariantAssociation
     with GQLManhattanAssociation
-    with GQLStudyLeadVariantAssociation {
+    with GQLStudyLeadVariantAssociation with GQLArguments {
 
-  val studyId =
-    Argument("studyId", StringType, description = "Study ID which links a top loci with a trait")
-
-  val geneId = Argument("geneId", StringType, description = "Gene ID using Ensembl identifier")
-
-  val phenotypeId = Argument(
-    "phenotypeId",
-    StringType,
-    description = "Phenotype ID using Ensembl identifier for the molecular traits")
-
-  val studyIds = Argument("studyIds", ListInputType(StringType), description = "List of study IDs")
-
-  val variantId = Argument(
-    "variantId",
-    StringType,
-    description = "Variant ID formated as CHR_POSITION_REFALLELE_ALT_ALLELE")
-
-  val variantIds = Argument(
-    "variantIds",
-    ListInputType(StringType),
-    description = "Variant ID formated as CHR_POSITION_REFALLELE_ALT_ALLELE")
-
-  val chromosome = Argument(
-    "chromosome",
-    StringType,
-    description = "Chromosome as String between 1..22 or X, Y, MT")
-
-  val pageIndex =
-    Argument("pageIndex", OptionInputType(IntType), description = "pagination index >= 0")
-
-  val pageSize = Argument("pageSize", OptionInputType(IntType), description = "pagination size > 0")
-
-  val dnaPosStart =
-    Argument("start", LongType, description = "Start position in a specified chromosome")
-
-  val dnaPosEnd = Argument("end", LongType, description = "End position in a specified chromosome")
-  val queryString = Argument("queryString", StringType, description = "Query text to search for")
-
-  val bioFeature = Argument(
-    "bioFeature",
-    StringType,
-    description = "BioFeature represents either a tissue, cell type, aggregation type, ...")
 
   val resolvers = DeferredResolver.fetchers(studiesFetcher, genesFetcher, variantsFetcher)
